# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Viral Waitlist & Referral Marketing Platform API
  description: |
    REST API for the Viral Waitlist & Referral Marketing Platform.

    ## Authentication

    All API requests require authentication via:
    - **API Key:** Header `X-API-Key: sk_live_xxxxx` or `Authorization: Bearer sk_live_xxxxx`
    - **JWT Token:** For dashboard users, `Authorization: Bearer <jwt>`

    API Key Format:
    - Test: `sk_test_` (test mode)
    - Live: `sk_live_` (production)

    ## Rate Limiting

    Rate limits by tier (per minute):
    - **Free:** 60 requests
    - **Starter:** 60 requests
    - **Pro:** 180 requests
    - **Enterprise:** 600+ requests (custom)

    Response headers include:
    - `X-RateLimit-Limit`: Maximum requests per window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when the rate limit resets
    - `X-Request-ID`: Unique request identifier for debugging

    ## Error Handling

    All error responses follow a consistent format with both a human-readable message and a machine-readable error code:

    ```json
    {
      "error": "User-friendly error message",
      "code": "ERROR_CODE"
    }
    ```

    Error codes use `UPPER_SNAKE_CASE` format for easy programmatic handling.

    ### Common Error Codes

    - `INVALID_INPUT` - Validation error or malformed request
    - `UNAUTHORIZED` - Missing or invalid authentication
    - `FORBIDDEN` - Insufficient permissions
...

...

...
servers:
...
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /api/v1/webhooks:
    post:
      tags:
        - Webhooks
      summary: Create webhook
      description: Create a new webhook endpoint
      operationId: createWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '201':
          description: Webhook created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  webhook:
                    $ref: '#/components/schemas/Webhook'
                  secret:
                    type: string
                    description: Webhook signing secret (only returned once)
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      tags:
        - Webhooks
      summary: List webhooks
      description: Retrieve all webhooks for the authenticated account
      operationId: listWebhooks
      parameters:
        - name: campaign_id
          in: query
          description: Filter by campaign ID
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of webhooks. Returns empty array [] if no webhooks found, never null.
          content:
            application/json:
              schema:
                type: array
                description: Array of webhooks
                items:
                  $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/webhooks/{webhook_id}:
    parameters:
      - $ref: '#/components/parameters/WebhookIdParam'

    get:
      tags:
        - Webhooks
      summary: Get webhook details
      description: Retrieve detailed information about a webhook
      operationId: getWebhook
      responses:
        '200':
          description: Webhook details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags:
        - Webhooks
      summary: Update webhook
      description: Update a webhook's configuration
      operationId: updateWebhook
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Webhook updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags:
        - Webhooks
      summary: Delete webhook
      description: Delete a webhook endpoint
      operationId: deleteWebhook
      responses:
        '204':
          description: Webhook deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/webhooks/{webhook_id}/test:
    post:
      tags:
        - Webhooks
      summary: Test webhook
      description: Send a test event to a webhook endpoint
      operationId: testWebhook
      parameters:
        - $ref: '#/components/parameters/WebhookIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                event_type:
                  type: string
                  description: Event type to send in test
                  default: "webhook.test"
      responses:
        '200':
          description: Test webhook sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  delivery:
                    $ref: '#/components/schemas/WebhookDelivery'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API Key authentication (sk_live_* or sk_test_*)

    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication for dashboard users
  parameters:
...
    WebhookIdParam:
      name: webhook_id
      in: path
      required: true
      description: Webhook unique identifier
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Bad request - Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            validation:
              value:
                error: "Invalid request format. Please check your JSON syntax."
                code: "INVALID_INPUT"
            missing_field:
              value:
                error: "Required field is missing"
                code: "INVALID_INPUT"
    Unauthorized:
      description: Unauthorized - Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Authorization token is missing or invalid"
            code: "UNAUTHORIZED"
...
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          examples:
            campaign:
              value:
                error: "Campaign not found"
                code: "CAMPAIGN_NOT_FOUND"
            user:
              value:
                error: "User not found"
                code: "USER_NOT_FOUND"
...
    TooManyRequests:
      description: Too many requests - Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          description: Maximum requests per window
          schema:
            type: integer
        X-RateLimit-Remaining:
          description: Remaining requests in current window
          schema:
            type: integer
        X-RateLimit-Reset:
          description: Unix timestamp when rate limit resets
          schema:
            type: integer
        Retry-After:
          description: Seconds to wait before retrying
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Rate limit exceeded. Please try again later."
            code: "RATE_LIMIT_EXCEEDED"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "An internal server error occurred. Please try again later."
            code: "INTERNAL_ERROR"
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: User-friendly error message
        code:
          type: string
          description: Machine-readable error code in UPPER_SNAKE_CASE format
          example: "INVALID_INPUT"
      example:
        error: "Invalid request format"
        code: "INVALID_INPUT"

    # ==================== CORE MODELS ====================
...
    Webhook:
      type: object
      required:
        - id
        - account_id
        - url
        - events
        - status
      properties:
        id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        campaign_id:
          type: string
          format: uuid
          description: Optional - campaign-specific webhook
        url:
          type: string
          format: uri
          maxLength: 500
        events:
          type: array
          items:
            type: string
            enum:
              - user.created
              - user.updated
              - user.verified
              - user.deleted
              - user.position_changed
              - user.converted
              - referral.created
              - referral.verified
              - referral.converted
              - reward.earned
              - reward.delivered
              - reward.redeemed
              - campaign.milestone
              - campaign.launched
              - campaign.completed
              - email.sent
              - email.delivered
              - email.opened
              - email.clicked
              - email.bounced
        status:
          type: string
          enum: [active, paused, failed]
        retry_enabled:
          type: boolean
          default: true
        max_retries:
          type: integer
          default: 5
        total_sent:
          type: integer
        total_failed:
          type: integer
        last_success_at:
          type: string
          format: date-time
        last_failure_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    WebhookDelivery:
      type: object
      required:
        - id
        - webhook_id
        - event_type
        - status
      properties:
        id:
          type: string
          format: uuid
        webhook_id:
          type: string
          format: uuid
        event_type:
          type: string
        payload:
          type: object
          additionalProperties: true
        status:
          type: string
          enum: [pending, success, failed]
        response_status:
          type: integer
          description: HTTP status code from webhook endpoint
        response_body:
          type: string
        duration_ms:
          type: integer
          description: Request duration in milliseconds
        attempt_number:
          type: integer
        next_retry_at:
          type: string
          format: date-time
        error_message:
          type: string
        created_at:
          type: string
          format: date-time
        delivered_at:
          type: string
          format: date-time

    # ==================== CONFIGURATION SCHEMAS ====================
...
    CreateWebhookRequest:
      type: object
      required:
        - url
        - events
      properties:
        url:
          type: string
          format: uri
          maxLength: 500
        campaign_id:
          type: string
          format: uuid
          description: Optional - make this campaign-specific
        events:
          type: array
          items:
            type: string
          minItems: 1
        retry_enabled:
          type: boolean
          default: true
        max_retries:
          type: integer
          default: 5
          minimum: 1
          maximum: 10
    UpdateWebhookRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
          maxLength: 500
        events:
          type: array
          items:
            type: string
          minItems: 1
        status:
          type: string
          enum: [active, paused]
        retry_enabled:
          type: boolean
        max_retries:
          type: integer
          minimum: 1
          maximum: 10

    # ==================== RESPONSE SCHEMAS ====================
...